More clang++ compile fixes.  All branches build now.
git-svn-id: svn://pqxx.org/libpqxx@1837 c3353b84-d008-0410-9ed9-b55c4f7fa7e8

+++ branches/3.0/test/test053.cxx
@@ -48,7 +48,7 @@ public:
   {
     char Buf[200];
     largeobjectaccess O(T, m_Object, ios::in);
-    const size_t len = O.read(Buf, sizeof(Buf)-1);
+    const size_t len = size_t(O.read(Buf, sizeof(Buf)-1));
     PQXX_CHECK_EQUAL(
 	string(Buf, len),
 	Contents,
