+++ branches/3.0/test/test055.cxx
@@ -32,7 +32,7 @@ public:
             "to large object #" << m_Object.id() << endl;
 
     char Buf[200];
-    const size_t len = A.read(Buf, sizeof(Buf)-1);
+    const size_t len = size_t(A.read(Buf, sizeof(Buf)-1));
     PQXX_CHECK_EQUAL(
 	string(Buf, len),
 	Contents,
